requirements.txt
pymongo==4.6.3
minio==7.2.7
requests==2.32.3
python-dotenv==1.0.1
pandas==2.2.2
rapidfuzz==3.9.6
matplotlib==3.9.2
